Complete Task #39: /courses page with two tabs and styled layout
