+++ src/styles/courses.css
@@ -77,3 +77,30 @@
   border-radius: 9999px;
   background-color: #52796f;
 }
+.tab-button {
+  background-color: #2f3e46;
+  color: #cad2c5;
+  border: none;
+  border-radius: 16px;
+  margin: 0 10px;
+  padding: 12px 90px;
+  font-weight: 600;
+  font-size: 1.1rem;
+  cursor: pointer;
+  transition:
+    background-color 0.3s,
+    color 0.3s,
+    transform 0.2s;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+}
+
+.tab-button:hover {
+  transform: translateY(-1px);
+  background-color: #3a5042;
+}
+
+.tab-button.active {
+  background-color: #cad2c5;
+  color: #2f3e46;
+  font-weight: 700;
+}
